requirements.txt
asgiref==3.7.2
bleach==6.1.0
Django==4.2.6
django-cors-headers==4.3.0
django-environ==0.11.2
django-summernote==0.8.20.0
django-tinymce==3.6.1
djangorestframework==3.14.0
ordered-set==4.1.0
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
webencodings==0.5.1
whitenoise==6.6.0
